change: Move the folder 'pages' to 'src'
